Filter out nested <pre>, <span>, and <code>.
This was causing problems with my markdown editing plugin. For inline backticks
it replaced them with `<code>`, but for multi-line backticks it replaced them
with
```
<pre>
   <code> ...</code>
</pre>
```

No combination of crayon's options would allow both the inline backticks and
multi-line backticks to display properly.

Because the plugin uses jetpack's markdown code and parser, this is likely a
systemic problem.